@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { LoginComponent } from './components/login/login.component';
@@ -61,75 +61,68 @@ import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
 import { DescargaInventarioComponent } from './components/descarga-inventario/descarga-inventario.component';
 
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    LoginComponent,
-    PerfilComponent,
-    NavbarComponent,
-    AddproductoComponent,
-    DashboardComponent,
-    AddsolicitudComponent,
-    AlertaComponent,
-    RedirectmodelComponent,
-    ProductosComponent,
-    StockComponent,
-    StockCriticoComponent,
-    DetalleProductoComponent,
-    AgregarProductoModalComponent,
-    DetalleStockComponent,
-    AgregarStockModalComponent,
-    AgregarLugarModalComponent,
-    LugarComponent,
-    EditProductModalComponent,
-    EditStockModalComponent,
-    EditLugarModalComponent,
-    DashboardSolicitudComponent,
-    DocumentoComponent,
-    DetalleSolicitudComponent,
-    SolicitudesComponent,
-    EditEstadoSolComponent,
-    CarritoComponent,
-    NavbarcartComponent,
-    SolicitudComponent,
-    MisolicitudComponent,
-    DescargaInventarioComponent
-  ],
-  imports: [
-    BrowserModule,
-    HttpClientModule,
-    AppRoutingModule,
-    BrowserAnimationsModule,
-    MatCardModule,
-    MatTabsModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatIconModule,
-    MatTableModule,
-    MatToolbarModule,
-    ReactiveFormsModule,
-    MatGridListModule,
-    FontAwesomeModule,
-    FormsModule,
-    MatDatepickerModule,
-    MatNativeDateModule,
-    MatListModule,
-    MatMomentDateModule,
-    MatSelectModule,
-    MatButtonModule,
-    MatCheckboxModule,
-    MatDialogModule,
-    BarcodeScannerLivestreamModule,
-    MatSidenavModule,
-    OAuthModule.forRoot(),
-    MatSidenavModule,
-    MatListModule,
-    CdkTableModule,
-    MatInputModule,
-    MatExpansionModule,
-    MatProgressSpinnerModule
-  ],
-  providers: [],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        LoginComponent,
+        PerfilComponent,
+        NavbarComponent,
+        AddproductoComponent,
+        DashboardComponent,
+        AddsolicitudComponent,
+        AlertaComponent,
+        RedirectmodelComponent,
+        ProductosComponent,
+        StockComponent,
+        StockCriticoComponent,
+        DetalleProductoComponent,
+        AgregarProductoModalComponent,
+        DetalleStockComponent,
+        AgregarStockModalComponent,
+        AgregarLugarModalComponent,
+        LugarComponent,
+        EditProductModalComponent,
+        EditStockModalComponent,
+        EditLugarModalComponent,
+        DashboardSolicitudComponent,
+        DocumentoComponent,
+        DetalleSolicitudComponent,
+        SolicitudesComponent,
+        EditEstadoSolComponent,
+        CarritoComponent,
+        NavbarcartComponent,
+        SolicitudComponent,
+        MisolicitudComponent,
+        DescargaInventarioComponent
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        AppRoutingModule,
+        BrowserAnimationsModule,
+        MatCardModule,
+        MatTabsModule,
+        MatFormFieldModule,
+        MatInputModule,
+        MatIconModule,
+        MatTableModule,
+        MatToolbarModule,
+        ReactiveFormsModule,
+        MatGridListModule,
+        FontAwesomeModule,
+        FormsModule,
+        MatDatepickerModule,
+        MatNativeDateModule,
+        MatListModule,
+        MatMomentDateModule,
+        MatSelectModule,
+        MatButtonModule,
+        MatCheckboxModule,
+        MatDialogModule,
+        BarcodeScannerLivestreamModule,
+        MatSidenavModule,
+        OAuthModule.forRoot(),
+        MatSidenavModule,
+        MatListModule,
+        CdkTableModule,
+        MatInputModule,
+        MatExpansionModule,
+        MatProgressSpinnerModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule { }
